Fixed glitch where you had to double tap to edit text view.
Here is why I think this was happening: there was a tapGestureRecognizer (being used to recognizer redactions) that ignored a single tap if it wasn't a redaction. This gesture therefore overrode the normal single tap to edit function. TextView also goes into editing mode when it is double tapped, so double taps still went through. The solution to this was to simply make the tapGestureRecognizer display the keyboard and select the text if it was in edit mode (ignore it if it was in redaction mode). Now I know if I want to stop something from performing the normal action in response to a gesture, to add a gesture recognizer to "override" it's normal behavior.